+++ imports/ui/Index.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+export default class Index extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <div>Hello!</div>
+    );
+  }
+}
